postcodes/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pika==1.2.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
requests==2.27.1
tomli==2.0.1
urllib3==1.26.9
yarl==1.7.2
